refactor: extract form input's into a seperate component.
