+++ src/features/user/Settings/EditProfile/AddressManagment/AddressManagement.module.scss
@@ -83,6 +83,9 @@
 // address action form
 
 .addressFormContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 40px;
   min-width: 565px;
   background: rgba(255, 255, 255, 1);
   border-radius: 16px;
@@ -91,7 +94,7 @@
   left: 50%;
   top: 50%;
   transform: translate(-50%, -50%);
-  h1 {
+  h2 {
     font-weight: 700;
     font-size: $fontSixteen;
   }
@@ -101,7 +104,6 @@
   display: flex;
   flex-direction: column;
   gap: 23px;
-  margin-top: 20px;
 }
 
 .formButtonContainer {
@@ -109,7 +111,6 @@
   gap: 8px;
   justify-content: space-between;
   align-items: end;
-  margin-top: 40px;
   .addAddressButton,
   .cancelButton {
     min-width: 252px;
